requirements/dev.txt
-r base.txt

django-debug-toolbar==2.2
ipdb==0.13.1
